// frontend/src/api/generated/types.gen.ts
export type SeasonRead = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description: string;
    /**
     * Is Running
     */
    is_running: boolean;
    /**
     * Created At
     */
    created_at: string;
};